+++ javascripts/index.js
@@ -1,7 +1,9 @@
 // Import our SCSS files so webpack will process them
 import '../stylesheets/index.scss'
 import displayPlatformSpecificContent from './display-platform-specific-content'
+import displayToolSpecificContent from './display-tool-specific-content'
 import explorer from './explorer'
+import scrollUp from './scroll-up'
 import search from './search'
 import nav from './nav'
 import browserDateFormatter from 'browser-date-formatter'
@@ -12,15 +14,20 @@ import localization from './localization'
 import helpfulness from './helpfulness'
 import experiment from './experiment'
 import copyCode from './copy-code'
-import { fillCsrf } from './get-csrf'
 import initializeEvents from './events'
 import filterCards from './filter-cards'
 import allArticles from './all-articles'
 import devToc from './dev-toc'
+import releaseNotes from './release-notes'
+import showMore from './show-more'
+import airgapLinks from './airgap-links'
+import toggleImages from './toggle-images'
 
 document.addEventListener('DOMContentLoaded', async () => {
   displayPlatformSpecificContent()
+  displayToolSpecificContent()
   explorer()
+  scrollUp()
   search()
   nav()
   browserDateFormatter()
@@ -32,8 +39,11 @@ document.addEventListener('DOMContentLoaded', async () => {
   filterCards()
   allArticles()
   devToc()
-  await fillCsrf() // this must complete before any POST calls
-  initializeEvents() // requires fillCsrf to complete
-  experiment() // requires fillCsrf to complete
-  helpfulness() // requires fillCsrf to complete
+  showMore()
+  airgapLinks()
+  releaseNotes()
+  initializeEvents()
+  experiment()
+  helpfulness()
+  toggleImages()
 })