+++ middleware/breadcrumbs.js
@@ -1,7 +1,9 @@
 const path = require('path')
 const { getPathWithoutLanguage } = require('../lib/path-utils')
+const nonEnterpriseDefaultVersion = require('../lib/non-enterprise-default-version')
+const removeFPTFromPath = require('../lib/remove-fpt-from-path')
 
-module.exports = async (req, res, next) => {
+module.exports = async function breadcrumbs (req, res, next) {
   if (!req.context.page) return next()
   if (req.context.page.hidden) return next()
 
@@ -16,15 +18,27 @@ module.exports = async (req, res, next) => {
   // drop first '/'
   pathParts.shift()
 
-  // drop the version segment so pathParts now starts with /product
-  pathParts.shift()
-
   const productPath = path.posix.join('/', req.context.currentProduct)
+
+  if (!req.context.siteTree[req.language][req.context.currentVersion].products) {
+    return next()
+  }
+
   const product = req.context.siteTree[req.language][req.context.currentVersion].products[req.context.currentProduct]
 
+  if (!product) {
+    return next()
+  }
+
   req.context.breadcrumbs.product = {
-    href: path.posix.join('/', req.context.currentVersion, productPath),
+    href: removeFPTFromPath(path.posix.join('/', req.context.currentLanguage, req.context.currentVersion, productPath)),
     title: product.title
+  }
+
+  // if this is not FPT, drop the version segment so pathParts now starts with /product
+  // if this is FPT, there is no version segment so pathParts already starts with /product
+  if (req.context.currentVersion !== nonEnterpriseDefaultVersion) {
+    pathParts.shift()
   }
 
   if (!pathParts[1]) return next()
@@ -32,7 +46,7 @@ module.exports = async (req, res, next) => {
   // get category path
   // e.g., `getting-started-with-github` in /free-pro-team@latest/github/getting-started-with-github
   // or /enterprise-server@2.21/github/getting-started-with-github
-  const categoryPath = path.posix.join('/', req.context.currentVersion, productPath, pathParts[1])
+  const categoryPath = removeFPTFromPath(path.posix.join('/', req.context.currentLanguage, req.context.currentVersion, productPath, pathParts[1]))
 
   const category = product.categories[categoryPath]
 
@@ -49,7 +63,7 @@ module.exports = async (req, res, next) => {
   // e.g., /github/getting-started-with-github/learning-about-github
   let maptopic
   if (req.context.page.mapTopic) {
-    const maptopicPath = path.posix.join(categoryPath, pathParts[2])
+    const maptopicPath = req.path
 
     maptopic = category.maptopics[maptopicPath]
 
@@ -60,9 +74,7 @@ module.exports = async (req, res, next) => {
       title: maptopic.shortTitle || maptopic.title
     }
   } else {
-    // get article path
-    // e.g., /github/getting-started-with-github/githubs-products
-    const articlePath = path.posix.join(categoryPath, pathParts[2])
+    const articlePath = req.path
 
     // find parent maptopic if one exists
     // some categories don't have maptopics, e.g. site-policy
@@ -77,17 +89,7 @@ module.exports = async (req, res, next) => {
       }
     }
 
-    let articleKey = '/' + req.language + articlePath
-    let articlePage = req.context.pages[articleKey]
-
-    // fall back to English if localized article does not exist
-    if (!articlePage && req.language !== 'en') {
-      articleKey = '/en' + articlePath
-      articlePage = req.context.pages[articleKey]
-    }
-
-    if (!articlePage) return next()
-
+    const articlePage = req.context.page
     const articleTitle = await articlePage.renderProp('shortTitle', req.context, { textOnly: true, encodeEntities: true })
 
     req.context.breadcrumbs.article = {
@@ -95,6 +97,5 @@ module.exports = async (req, res, next) => {
       title: articleTitle
     }
   }
-
   return next()
 }