+++ middleware/contextualizers/rest.js
@@ -1,16 +1,18 @@
+const path = require('path')
 const rest = require('../../lib/rest')
-const { getVersionedPathWithLanguage } = require('../../lib/path-utils')
+const removeFPTFromPath = require('../../lib/remove-fpt-from-path')
 
-module.exports = async function (req, res, next) {
+module.exports = function restContext (req, res, next) {
   req.context.rest = rest
 
   // link to include in `Works with GitHub Apps` notes
   // e.g. /ja/rest/reference/apps or /en/enterprise/2.20/user/rest/reference/apps
-  req.context.restGitHubAppsLink = getVersionedPathWithLanguage(
-    '/developers/apps',
+  req.context.restGitHubAppsLink = removeFPTFromPath(path.join(
+    '/',
+    req.context.currentLanguage,
     req.context.currentVersion,
-    req.context.currentLanguage
-  )
+    '/developers/apps'
+  ))
 
   // ignore requests to non-REST reference paths
   if (!req.path.includes('rest/reference')) return next()