+++ tests/routing/redirects.js
@@ -1,7 +1,7 @@
 const path = require('path')
 const { isPlainObject } = require('lodash')
 const supertest = require('supertest')
-const app = require('../../server')
+const app = require('../../lib/app')
 const enterpriseServerReleases = require('../../lib/enterprise-server-releases')
 const nonEnterpriseDefaultVersion = require('../../lib/non-enterprise-default-version')
 const Page = require('../../lib/page')
@@ -23,6 +23,7 @@ describe('redirects', () => {
       basePath: path.join(__dirname, '../../content'),
       languageCode: 'en'
     })
+    page.buildRedirects()
     expect(isPlainObject(page.redirects)).toBe(true)
   })
 
@@ -32,10 +33,13 @@ describe('redirects', () => {
       basePath: path.join(__dirname, '../../content'),
       languageCode: 'en'
     })
-    expect(page.redirects['/articles']).toBe(`/en/${nonEnterpriseDefaultVersion}/github`)
-    expect(page.redirects['/en/articles']).toBe(`/en/${nonEnterpriseDefaultVersion}/github`)
-    expect(page.redirects['/common-issues-and-questions']).toBe(`/en/${nonEnterpriseDefaultVersion}/github`)
-    expect(page.redirects['/en/common-issues-and-questions']).toBe(`/en/${nonEnterpriseDefaultVersion}/github`)
+    page.buildRedirects()
+    expect(page.redirects[`/en/${nonEnterpriseDefaultVersion}/github`]).toBe('/en/github')
+    expect(page.redirects['/articles']).toBe('/en/github')
+    expect(page.redirects['/en/articles']).toBe('/en/github')
+    expect(page.redirects[`/en/${nonEnterpriseDefaultVersion}/articles`]).toBe('/en/github')
+    expect(page.redirects['/common-issues-and-questions']).toBe('/en/github')
+    expect(page.redirects['/en/common-issues-and-questions']).toBe('/en/github')
     expect(page.redirects[`/en/enterprise/${enterpriseServerReleases.latest}/user/articles`]).toBe(`/en/enterprise-server@${enterpriseServerReleases.latest}/github`)
     expect(page.redirects[`/en/enterprise/${enterpriseServerReleases.latest}/user/common-issues-and-questions`]).toBe(`/en/enterprise-server@${enterpriseServerReleases.latest}/github`)
   })
@@ -46,7 +50,8 @@ describe('redirects', () => {
       basePath: path.join(__dirname, '../../content'),
       languageCode: 'en'
     })
-    const expected = `/en/${nonEnterpriseDefaultVersion}/github/collaborating-with-issues-and-pull-requests/about-conversations-on-github`
+    page.buildRedirects()
+    const expected = '/en/github/collaborating-with-issues-and-pull-requests/about-conversations-on-github'
     expect(page.redirects['/en/articles/about-discussions-in-issues-and-pull-requests']).toBe(expected)
   })
 
@@ -122,7 +127,7 @@ describe('redirects', () => {
     test('redirect_from for renamed pages', async () => {
       const { res } = await get('/ja/desktop/contributing-to-projects/changing-a-remote-s-url-from-github-desktop')
       expect(res.statusCode).toBe(301)
-      const expected = `/ja/${nonEnterpriseDefaultVersion}/desktop/contributing-and-collaborating-using-github-desktop/changing-a-remotes-url-from-github-desktop`
+      const expected = '/ja/desktop/contributing-and-collaborating-using-github-desktop/changing-a-remotes-url-from-github-desktop'
       expect(res.headers.location).toBe(expected)
     })
   })
@@ -153,6 +158,12 @@ describe('redirects', () => {
       const res = await get('/ja/enterprise')
       expect(res.statusCode).toBe(301)
       expect(res.headers.location).toBe(japaneseEnterpriseHome)
+    })
+
+    test('hardcoded @latest redirects to latest version', async () => {
+      const res = await get('/en/enterprise-server@latest')
+      expect(res.statusCode).toBe(301)
+      expect(res.headers.location).toBe(enterpriseHome)
     })
   })
 
@@ -329,7 +340,7 @@ describe('redirects', () => {
   })
 
   describe('desktop guide', () => {
-    const desktopGuide = `/en/${nonEnterpriseDefaultVersion}/desktop/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request`
+    const desktopGuide = '/en/desktop/contributing-and-collaborating-using-github-desktop/creating-an-issue-or-pull-request'
     const japaneseDesktopGuides = desktopGuide.replace('/en/', '/ja/')
 
     test('no language code redirects to english', async () => {