+++ tests/unit/page.js
@@ -5,6 +5,12 @@ const prerenderedObjects = require('../../lib/graphql/static/prerendered-objects
 const allVersions = require('../../lib/all-versions')
 const enterpriseServerReleases = require('../../lib/enterprise-server-releases')
 const nonEnterpriseDefaultVersion = require('../../lib/non-enterprise-default-version')
+const { latest } = require('../../lib/enterprise-server-releases')
+const enterpriseServerVersions = Object.keys(allVersions).filter(v => v.startsWith('enterprise-server@'))
+
+const getLinkData = require('../../lib/get-link-data')
+jest.mock('../../lib/get-link-data')
+
 // get the `free-pro-team` segment of `free-pro-team@latest`
 const nonEnterpriseDefaultPlan = nonEnterpriseDefaultVersion.split('@')[0]
 
@@ -82,8 +88,8 @@ describe('Page class', () => {
     test('rewrites links to include the current language prefix and version', async () => {
       const page = await Page.init(opts)
       const context = {
-        page: { version: nonEnterpriseDefaultVersion },
-        currentVersion: nonEnterpriseDefaultVersion,
+        page: { version: `enterprise-server@${enterpriseServerReleases.latest}` },
+        currentVersion: `enterprise-server@${enterpriseServerReleases.latest}`,
         currentPath: '/en/github/collaborating-with-issues-and-pull-requests/about-branches',
         currentLanguage: 'en'
       }
@@ -92,14 +98,14 @@ describe('Page class', () => {
       expect(page.markdown.includes('(/articles/about-pull-requests)')).toBe(true)
       expect(page.markdown.includes('(/en/articles/about-pull-requests)')).toBe(false)
       expect($('a[href="/articles/about-pull-requests"]').length).toBe(0)
-      expect($(`a[href="/en/${nonEnterpriseDefaultVersion}/articles/about-pull-requests"]`).length).toBeGreaterThan(0)
+      expect($(`a[href="/en/${`enterprise-server@${enterpriseServerReleases.latest}`}/articles/about-pull-requests"]`).length).toBeGreaterThan(0)
     })
 
     test('rewrites links on prerendered GraphQL page include the current language prefix and version', async () => {
-      const graphqlVersion = allVersions[nonEnterpriseDefaultVersion].miscVersionName
+      const graphqlVersion = allVersions[`enterprise-server@${enterpriseServerReleases.latest}`].miscVersionName
       const $ = cheerio.load(prerenderedObjects[graphqlVersion].html)
       expect($('a[href^="/graphql/reference/input-objects"]').length).toBe(0)
-      expect($(`a[href^="/en/${nonEnterpriseDefaultVersion}/graphql/reference/input-objects"]`).length).toBeGreaterThan(0)
+      expect($(`a[href^="/en/enterprise-server@${enterpriseServerReleases.latest}/graphql/reference/input-objects"]`).length).toBeGreaterThan(0)
     })
 
     test('rewrites links in the intro to include the current language prefix and version', async () => {
@@ -114,7 +120,7 @@ describe('Page class', () => {
       await page.render(context)
       const $ = cheerio.load(page.intro)
       expect($('a[href="/articles/about-pull-requests"]').length).toBe(0)
-      expect($(`a[href="/en/${nonEnterpriseDefaultVersion}/articles/about-pull-requests"]`).length).toBeGreaterThan(0)
+      expect($('a[href="/en/articles/about-pull-requests"]').length).toBeGreaterThan(0)
     })
 
     test('does not rewrite links that include deprecated enterprise release numbers', async () => {
@@ -159,11 +165,11 @@ describe('Page class', () => {
         basePath: path.join(__dirname, '../fixtures'),
         languageCode: 'en'
       })
-      // set version to the latest enteprise version
+      // set version to the latest enterprise version
       const context = {
         currentVersion: `enterprise-server@${enterpriseServerReleases.latest}`,
         currentLanguage: 'en',
-        enterpriseServerVersions: Object.keys(allVersions).filter(id => id.startsWith('enterprise-server@'))
+        enterpriseServerVersions
       }
       let rendered = await page.render(context)
       let $ = cheerio.load(rendered)
@@ -185,6 +191,36 @@ describe('Page class', () => {
       $ = cheerio.load(rendered)
       expect($.text()).not.toBe('This text should render on any actively supported version of Enterprise Server')
       expect($.text()).toBe('This text should only render on non-Enterprise')
+    })
+
+    test('support next to-be-released Enterprise Server version in frontmatter', async () => {
+      // This fixture has `enterprise-server: '>=3.1'` hardcoded in the frontmatter
+      const page = await Page.init({
+        relativePath: 'page-versioned-for-next-enterprise-release.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en'
+      })
+      // set version to 3.0
+      const context = {
+        currentVersion: 'enterprise-server@3.0',
+        currentLanguage: 'en'
+      }
+      await expect(() => { return page.render(context) }).not.toThrow()
+    })
+
+    test('support next GitHub AE version in frontmatter', async () => {
+      // This fixture has `github-ae: 'next'` hardcoded in the frontmatter
+      const page = await Page.init({
+        relativePath: 'page-versioned-for-ghae-next.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en'
+      })
+      // set version to @latest
+      const context = {
+        currentVersion: 'github-ae@latest',
+        currentLanguage: 'en'
+      }
+      await expect(() => { return page.render(context) }).not.toThrow()
     })
   })
 
@@ -231,7 +267,7 @@ describe('Page class', () => {
 
     test('sets versioned values', async () => {
       const page = await Page.init(opts)
-      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe(`/en/${nonEnterpriseDefaultVersion}/github/collaborating-with-issues-and-pull-requests/about-branches`)
+      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe('/en/github/collaborating-with-issues-and-pull-requests/about-branches')
       expect(page.permalinks.find(permalink => permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.oldestSupported}`).href).toBe(`/en/enterprise-server@${enterpriseServerReleases.oldestSupported}/github/collaborating-with-issues-and-pull-requests/about-branches`)
     })
 
@@ -241,9 +277,8 @@ describe('Page class', () => {
         basePath: path.join(__dirname, '../../content'),
         languageCode: 'en'
       })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe(`/en/${nonEnterpriseDefaultVersion}`)
-      expect(page.permalinks.find(permalink => permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.oldestSupported}`).href).toBe(`/en/enterprise-server@${enterpriseServerReleases.oldestSupported}`)
       expect(page.permalinks.find(permalink => permalink.pageVersion === 'homepage').href).toBe('/en')
+      expect(page.permalinks.find(permalink => permalink.pageVersion === `enterprise-server@${enterpriseServerReleases.oldestSupported}`).href).toBe(`/en/enterprise-server@${enterpriseServerReleases.oldestSupported}`)
     })
 
     test('permalinks for dotcom-only pages', async () => {
@@ -252,7 +287,7 @@ describe('Page class', () => {
         basePath: path.join(__dirname, '../../content'),
         languageCode: 'en'
       })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe(`/en/${nonEnterpriseDefaultVersion}/github/getting-started-with-github/signing-up-for-a-new-github-account`)
+      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe('/en/github/getting-started-with-github/signing-up-for-a-new-github-account')
       expect(page.permalinks.length).toBe(1)
     })
 
@@ -274,7 +309,7 @@ describe('Page class', () => {
         basePath: path.join(__dirname, '../fixtures'),
         languageCode: 'en'
       })
-      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe(`/en/${nonEnterpriseDefaultVersion}/products/actions/some-category/some-article`)
+      expect(page.permalinks.find(permalink => permalink.pageVersion === nonEnterpriseDefaultVersion).href).toBe('/en/products/actions/some-category/some-article')
       expect(page.permalinks.length).toBe(1)
     })
 
@@ -288,6 +323,133 @@ describe('Page class', () => {
       const pageVersions = page.permalinks.map(permalink => permalink.pageVersion)
       expect(pageVersions.length).toBeGreaterThan(1)
       expect(pageVersions.includes(nonEnterpriseDefaultVersion)).toBe(false)
+    })
+  })
+
+  describe('learning tracks', () => {
+    let page
+
+    beforeEach(async () => {
+      page = await Page.init({
+        relativePath: 'article-with-learning-tracks.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en'
+      })
+    })
+
+    it('includes learning tracks specified in frontmatter', async () => {
+      expect(page.learningTracks).toStrictEqual([
+        'track_1',
+        'track_2',
+        'non_existing_track',
+        '{% if currentVersion == "free-pro-team@latest" %}dotcom_only_track{% endif %}',
+        '{% if currentVersion != "free-pro-team@latest" %}enterprise_only_track{% endif %}'
+      ])
+    })
+
+    it('renders learning tracks that have been defined', async () => {
+      getLinkData.mockImplementation((guides) => { return guides })
+      const guides = ['/path/guide1', '/path/guide2']
+      const context = {
+        currentLanguage: 'en',
+        currentProduct: 'snowbird',
+        currentVersion: nonEnterpriseDefaultVersion,
+        enterpriseServerVersions,
+        site: {
+          data: {
+            'learning-tracks': {
+              snowbird: {
+                track_1: {
+                  title: 'title',
+                  description: 'description',
+                  guides,
+                  featured_track: '{% if currentVersion == "free-pro-team@latest" %}true{% else %}false{% endif %}'
+                },
+                track_2: {
+                  title: 'title',
+                  description: 'description',
+                  guides,
+                  featured_track: '{% if enterpriseServerVersions contains currentVersion %}true{% else %}false{% endif %}'
+                },
+                dotcom_only_track: {
+                  title: 'title',
+                  description: 'description',
+                  guides
+                },
+                enterprise_only_track: {
+                  title: 'title',
+                  description: 'description',
+                  guides
+                }
+              }
+            }
+          }
+        }
+      }
+      // Test that Liquid versioning is respected during rendering.
+      // Start with Dotcom.
+      await page.render(context)
+      // To actually render the guides in this test, we would have to load context.pages and context.redirects;
+      // To avoid that we can just test that the function was called with the expected data.
+      expect(getLinkData).toHaveBeenCalledWith(guides, context)
+      // Tracks for dotcom should exclude enterprise_only_track and the featured track_1.
+      expect(page.learningTracks).toHaveLength(2)
+      const dotcomTrackNames = page.learningTracks.map(t => t.trackName)
+      expect(dotcomTrackNames.includes('track_2')).toBe(true)
+      expect(dotcomTrackNames.includes('dotcom_only_track')).toBe(true)
+      expect(page.featuredTrack.trackName === 'track_1').toBeTruthy()
+      expect(page.featuredTrack.trackName === 'track_2').toBeFalsy()
+
+      // Switch to Enterprise.
+      context.currentVersion = `enterprise-server@${latest}`
+      await page.render(context)
+      // Tracks for enterprise should exclude dotcom_only_track and the featured track_2.
+      expect(page.learningTracks).toHaveLength(2)
+      const ghesTrackNames = page.learningTracks.map(t => t.trackName)
+      expect(ghesTrackNames.includes('track_1')).toBe(true)
+      expect(ghesTrackNames.includes('enterprise_only_track')).toBe(true)
+      expect(page.featuredTrack.trackName === 'track_1').toBeFalsy()
+      expect(page.featuredTrack.trackName === 'track_2').toBeTruthy()
+    })
+  })
+
+  describe('includeGuides', () => {
+    let page
+
+    beforeEach(async () => {
+      page = await Page.init({
+        relativePath: 'article-with-includeGuides.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en'
+      })
+    })
+
+    it('includes guide paths specified in frontmatter', async () => {
+      expect(page.includeGuides).toStrictEqual(['/path/guide1', '/path/guide2', '/path/guide3'])
+    })
+
+    it('renders guides and topics', async () => {
+      getLinkData.mockImplementation(() => {
+        return [{
+          page: { topics: ['Spring', 'Summer'] }
+        }, {
+          page: { topics: ['Summer', 'Fall'] }
+        }, {
+          page: { topics: ['Fall', 'Winter'] }
+        }]
+      })
+      const guides = ['/path/guide1', '/path/guide2', '/path/guide3']
+      const context = {
+        currentVersion: nonEnterpriseDefaultVersion,
+        currentLanguage: 'en'
+      }
+      await page.render(context)
+      expect(getLinkData).toHaveBeenCalledWith(guides, context)
+      expect(page.includeGuides).toHaveLength(3)
+      expect(page.allTopics).toHaveLength(4)
+      expect(page.allTopics).toEqual(
+        expect.arrayContaining(['Spring', 'Summer', 'Fall', 'Winter'])
+      )
     })
   })
 
@@ -366,10 +528,34 @@ describe('Page class', () => {
       expect(page.versions['enterprise-server']).toBe('*')
     })
   })
+
+  describe('platform specific content', () => {
+    test('page.defaultPlatform frontmatter', async () => {
+      const page = await Page.init({
+        relativePath: 'actions/hosting-your-own-runners/configuring-the-self-hosted-runner-application-as-a-service.md',
+        basePath: path.join(__dirname, '../../content'),
+        languageCode: 'en'
+      })
+      expect(page.defaultPlatform).toBeDefined()
+      expect(page.defaultPlatform).toBe('linux')
+    })
+  })
+
+  describe('tool specific content', () => {
+    test('page.defaultTool frontmatter', async () => {
+      const page = await Page.init({
+        relativePath: 'default-tool.md',
+        basePath: path.join(__dirname, '../fixtures'),
+        languageCode: 'en'
+      })
+      expect(page.defaultTool).toBeDefined()
+      expect(page.defaultTool).toBe('cli')
+    })
+  })
 })
 
 describe('catches errors thrown in Page class', () => {
-  test('frontmatter parsing error', () => {
+  test('frontmatter parsing error', async () => {
     async function getPage () {
       return await Page.init({
         relativePath: 'page-with-frontmatter-error.md',
@@ -378,10 +564,10 @@ describe('catches errors thrown in Page class', () => {
       })
     }
 
-    expect(getPage).rejects.toThrowError('invalid frontmatter entry')
+    await expect(getPage).rejects.toThrowError('invalid frontmatter entry')
   })
 
-  test('missing versions frontmatter', () => {
+  test('missing versions frontmatter', async () => {
     async function getPage () {
       return await Page.init({
         relativePath: 'page-with-missing-product-versions.md',
@@ -390,10 +576,10 @@ describe('catches errors thrown in Page class', () => {
       })
     }
 
-    expect(getPage).rejects.toThrowError('versions')
+    await expect(getPage).rejects.toThrowError('versions')
   })
 
-  test('English page with a version in frontmatter that its parent product is not available in', () => {
+  test('English page with a version in frontmatter that its parent product is not available in', async () => {
     async function getPage () {
       return await Page.init({
         relativePath: 'admin/some-category/some-article-with-mismatched-versions-frontmatter.md',
@@ -405,7 +591,7 @@ describe('catches errors thrown in Page class', () => {
     expect(getPage).rejects.toThrowError(/`versions` frontmatter.*? product is not available in/)
   })
 
-  test('non-English page with a version in frontmatter that its parent product is not available in', () => {
+  test('non-English page with a version in frontmatter that its parent product is not available in', async () => {
     async function getPage () {
       return await Page.init({
         relativePath: 'admin/some-category/some-article-with-mismatched-versions-frontmatter.md',
@@ -414,6 +600,6 @@ describe('catches errors thrown in Page class', () => {
       })
     }
 
-    expect(getPage).rejects.toThrowError(/`versions` frontmatter.*? product is not available in/)
+    await expect(getPage).rejects.toThrowError(/`versions` frontmatter.*? product is not available in/)
   })
 })