+++ webpack.config.js
@@ -1,17 +1,19 @@
 const path = require('path')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const CopyWebpackPlugin = require('copy-webpack-plugin')
-const { EnvironmentPlugin } = require('webpack')
+const { EnvironmentPlugin, ProvidePlugin } = require('webpack')
 const { reactBabelOptions } = require('./lib/react/babel')
 
 module.exports = {
-  devtool: 'source-map', // this prevents webpack from using eval
+  mode: 'development',
+  devtool: process.env.NODE_ENV === 'development' ? 'eval' : 'source-map', // no 'eval' outside of development
   entry: './javascripts/index.js',
   output: {
     filename: 'index.js',
     path: path.resolve(__dirname, 'dist'),
     publicPath: '/dist'
   },
+  stats: 'errors-only',
   module: {
     rules: [
       {
@@ -85,13 +87,12 @@ module.exports = {
         { from: 'node_modules/@primer/css/fonts', to: 'fonts' }
       ]
     }),
-    new EnvironmentPlugin(['NODE_ENV'])
-  ],
-  resolve: {
-    alias: {
-      // Hogan uses `new Function` which breaks content security policy
-      // Turns out, we aren't even using it anyways!
-      'hogan.js': path.resolve(__dirname, 'javascripts/fake-hogan.js')
-    }
-  }
+    new EnvironmentPlugin({
+      NODE_ENV: 'development', // use 'development' unless process.env.NODE_ENV is defined
+      DEBUG: false
+    }),
+    new ProvidePlugin({
+      process: 'process/browser'
+    })
+  ]
 }