+++ translations/ja-JP/content/admin/release-notes.md
@@ -1,8 +1,10 @@
 ---
 title: リリースノート
-intro: The release notes for {{ allVersions[currentVersion].versionTitle }}.
+intro: '{{ allVersions[currentVersion].versionTitle }} のリリースノートです。'
 layout: release-notes
 versions:
   enterprise-server: '*'
+topics:
+  - Enterprise
 ---